Add back autoindex styling with more specific location
The original problem with this XSLT styling was that the examples online
have you apply the styling to "location /" which means that Nginx is trying
to apply this styling to every asset served, be it the generated output of
"autoindex_format xml" or anything else that LOOKS like an XML file.

To correct this, we only enable autoindex and XSLT styling for paths matching
the regex "/$" so it only applies to paths ending in a forward slash, which is
definitely only going to be directories and not repo assests.