Setup certbot / tls inline now
This will check the system, see if the nginx letsencrypt options are
present and the certificate "file" (it's a symlink) and if either is
missing go ahead and run certbot.  This may need to ignore errors from
certbot if we are doing an initial setup locally and shipping it off to
where certbot will actually work.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>